# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Animal Spotting API",
    "description": "JSON REST API for tracking animal sightings"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "basePath": "",
  "paths": {
...
    "/sightings/{page}": {
      "get": {
        "summary": "Get a collection of sightings in pages",
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/sightings/{id}": {
      "get": {
        "summary": "Get a specific sighting by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/sightings/": {
      "post": {
        "summary": "Create a specific sighting",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "example": "any"
                  },
                  "trailName": {
                    "example": "any"
                  },
                  "animalName": {
                    "example": "any"
                  },
                  "animalCount": {
                    "example": "any"
                  },
                  "animalSpecies": {
                    "example": "any"
                  },
                  "sightingTime": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a specific sighting by id",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthenticationKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-AUTH-KEY"
      }
    }
  },
  "security": [
    {
      "AuthenticationKey": []
    }
  ]
}